rename start file to connect
